Hotfix for bug causing worker launches to fail.
Problem Description: Initially, when the manager received a request to start a worker from a stopped state, it would first check
to verify that the notebook server is not running on the worker before starting the notebook server. After looking
at the logs, it was found that, in some fraction of the time, the verification call - a call to the is_noteboook_running function -
would go to the wrong worker, a worker that has the notebook server already running. In this case, the manager would
erroneously think that the notebook server was already up and running on the worker to be started. Consequently, the manager
would fail to start the notebook server on the worker resuling in an unsuccessful launch. On the browser, the user would fail
to see the expected notebook interface. They would just see a red progress bar with a message indicating that the
spawn failed. To move forward, the user would then have to click on the Home button then the Start My Server button
to retry the launch.

Solution: This commit introduces a few changes to reduce the rate of unsuccessful launches, which at the moment is roughly 50%
for a cluster with more than about 30 worker instances already running. Namely:

- In the start_worker_server function, removing the verification call that checks whether the notebook server is running, if the worker is being started from a stopped state. The verification call is unnecessary given that the notebook server would not be running in the first place. Thus, the change here is to have the manager run the notebook server start command without doing the verification call. I note that this is a workaround to the problem
described. Namely, instead of taking the risk to make the verification call to the wrong worker, the manager just doesn't
make the call in the first place.

- Adding a while loop to the remote_notebook_start function that retries the notebook server start command until either
the notebook server starts on the worker or the maximum number of retries is reached. Initially, the notebook server
start command would be called only once. By adding retries, the probability that the launch will be successful increases.

- Adding more conditions to the logic in the is_notebook_running function to ensure that the function returns True only if
the notebook server is running on the right worker instance and the worker instance belongs to the user in question.
The logs showed that, in some fraction of the time, the function returns True after verifying that the notebook server is
running on the wrong instance. The logic, thus, needed tightening.

- Adding more debugging statements for the sake of ease of debugging when grepping through the manager logs.

I have tested the above changes in a cluster of about 60 worker instances wherein I launched each worker instance sequentially
and had them all running simultaneously. My observation was that the average time to launch reduced by 40% - from approx 1 min 20 sec
to about 50 min. In addition, no worker launched unsuccessfully, except for 1 that had a different issue going on.

Refactor changes introduced to reduce unsuccessful-launch-rate

Removed unnecessary try/except blocks. Reduced number of attempts in each call to the is_notebook_running function from 30 to 10.
Increased gen.sleep time in is_notebook_running function from 1 to 3 seconds. The reasoning here is that 3 seconds should be enough
for the notebook server to start.