Upgrade from Fabric3 to Fabric.
Fabric3 is an unauthorized fork of the main fabric project. It was
originally forked to add python3 support, but the main fabric project
now supports python3 and also has gone through a significant rewrite.

The motivation for upgrading, besides the fact that Fabric3 is
deprecated and no longer supported, is the fact that fabric is
thread-safe, providing better support for concurrency.